% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_log.R
\name{add_log_entry}
\alias{add_log_entry}
\title{Add a log entry}
\usage{
add_log_entry(
  r,
  category = character(),
  name = character(),
  value = character()
)
}
\arguments{
\item{r}{Shiny reactive value used to communicate between modules}

\item{category}{Category of the log entry, eg : "SQL query" (character)}

\item{name}{Name of the log entry, eg : "Add a new dataset" (character)}

\item{value}{Value of the log entry, eg : "INSERT INTO ..."}
}
\description{
Add a log entry
}
\examples{
\dontrun{
add_log_entry(r = r, category = "delete_data", name = "SQL query", value = "DELETE FROM datasets WHERE id = 5")
}
}
