% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_persons_to_subset}
\alias{add_persons_to_subset}
\title{Add patients to a subset}
\usage{
add_persons_to_subset(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  persons = tibble::tibble(),
  subset_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{m}{A shiny::reactiveValues object, used to communicate between modules}

\item{persons}{data variable containing patients / persons (data.frame / tibble)}

\item{subset_id}{ID of the subset (integer)}

\item{i18n}{Translator object from shiny.i18n library}

\item{ns}{Shiny namespace}
}
\description{
Add patients to a subset
}
\examples{
\dontrun{
persons <- tibble::tribble(~person_id, 123L, 456L, 789L)
subset_add_patients(output = output, r = r, m = m, persons = persons, subset_id = 3, i18n = i18n, ns = ns)
}
}
