% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_thesaurus_item}
\alias{add_thesaurus_item}
\title{Add thesaurus item}
\usage{
add_thesaurus_item(
  output,
  r = shiny::reactiveValues(),
  thesaurus_name = character(),
  item_name = character(),
  item_unit = NA_character_,
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between modules}

\item{thesaurus_name}{Name of the thesaurus where the item will be added (character)}

\item{item_name}{Name of the item (character)}

\item{item_unit}{Unit of the item (character)}

\item{i18n}{shiny.i18n object for translations}

\item{ns}{shiny namespace object, used to render messages on message bars}
}
\description{
Add an item to a thesaurus
}
\details{
To add an item in a thesaurus, you need to specify its name and unit.
The item ID will be the ID of the last item of this thesaurus, plus one.
}
\examples{
\dontrun{
# Create item "Heart rate" in thesaurus "MIMIC-IV"
add_thesaurus_item(output = output, r = r, thesaurus_name = "MIMIC-IV", item_name = "Heart rate", item_unit = "bpm")
}
}
