% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_vocabulary_concept}
\alias{add_vocabulary_concept}
\title{Add vocabulary concept}
\usage{
add_vocabulary_concept(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  vocabulary_id = character(),
  concept_name = character(),
  domain_id = character(),
  concept_class_id = character(),
  concept_code = character(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{m}{A shiny::reactiveValues object, used to communicate between modules}

\item{vocabulary_id}{Value of vocabulary_id where the concept will be added (character)}

\item{concept_name}{Value of concept_name (character)}

\item{domain_id}{Value of domain_id, must be present in database (character)}

\item{concept_class_id}{Value of concept_class_id, must be present in database (character)}

\item{concept_code}{Value of concept_code (character)}

\item{i18n}{Translator object from shiny.i18n library}

\item{ns}{Shiny namespace}
}
\description{
Add a new concept in the database concept table
}
\details{
See \href{https://ohdsi.github.io/CommonDataModel/cdm60.html#CONCEPT}{OMOP concept table} for more details.\cr\cr
Vocabulary_id must be present in the database when you add a new concept.\cr
Not null values are required for concept_name, domain_id, concept_class_id and concept_code.\cr
The value you provide for domain_id & concept_class_id must be present in the database.
}
\examples{
\dontrun{

}
}
