% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdwtools.R
\name{cdwtools}
\alias{cdwtools}
\title{Run the Shiny Application}
\usage{
cdwtools(options = list(), language = "EN", ...)
}
\arguments{
\item{options}{A list of options that could be passed to shiny::shinyApp() function (list)}

\item{language}{Default language to use in the App (character)}

\item{...}{arguments to pass to golem_opts.}
}
\description{
Runs the cdwtools Shiny Application.\cr
Use language argument to choose default language to use ("EN" or "FR").\cr
Use options to set options to shiny::shinyApp() function, see documentation of shinyApp for more informations.
}
\examples{
\dontrun{
cdwtools(language = "EN")
}
See `?golem::get_golem_options` for more details.
}
