% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{check_authentification}
\alias{check_authentification}
\title{Check user authentification}
\usage{
check_authentification(db)
}
\arguments{
\item{db}{A DBI database connection object}
}
\description{
Check user authentification using shinymanager library
}
