% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{coalesce2}
\alias{coalesce2}
\title{Coalesce2}
\usage{
coalesce2(type, x)
}
\arguments{
\item{type}{Type of the variable c("char", "int") (character)}

\item{x}{Variable}
}
\value{
Return NA if variable is null or empty, returns the variable if not null & not empty
}
\description{
Returns a NA value
}
\examples{
coalesce2("char", "my_char")
}
