% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{create_datatable_cache}
\alias{create_datatable_cache}
\title{Create cache for datatable data}
\usage{
create_datatable_cache(
  output,
  r = shiny::reactiveValues(),
  i18n = character(),
  module_id = character(),
  ids = integer(),
  category = character()
)
}
\arguments{
\item{output}{Shiny output value, to show message bars}

\item{r}{Shiny reactive value, to communicate between modules}

\item{i18n}{Translator object from shiny.i18n library}

\item{module_id}{ID of current page / module (character)}

\item{ids}{IDs of rows in database}

\item{category}{Category of cache, depending of the page of Settings (character)}
}
\description{
Create cache for datatable data
}
