% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{create_scripts_thesaurus}
\alias{create_scripts_thesaurus}
\title{Create scripts thesaurus}
\usage{
create_scripts_thesaurus(
  output,
  r = shiny::reactiveValues(),
  data_source_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{output}

\item{r}{r}

\item{data_source_id}{data_source_id}

\item{i18n}{i18n}

\item{ns}{ns}
}
\description{
Create scripts thesaurus
}
