% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{db_create_tables}
\alias{db_create_tables}
\title{Create tables}
\usage{
db_create_tables(db)
}
\arguments{
\item{db}{Database connection object (DBI)}
}
\description{
Create all application database required tables. It creates the tables if they do not already exist.
}
