% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{delete_element}
\alias{delete_element}
\title{Delete element}
\usage{
delete_element(
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  session,
  input,
  output,
  ns = character(),
  i18n = character(),
  delete_prefix = character(),
  dialog_title = character(),
  dialog_subtext = character(),
  react_variable = character(),
  table = character(),
  r_table = character(),
  id_var_sql = character(),
  id_var_r = character(),
  delete_message = character(),
  reload_variable = character(),
  information_variable = character(),
  translation = TRUE,
  app_folder = character(),
  prefix = character(),
  r_message_bar = FALSE
)
}
\arguments{
\item{r}{Shiny reactive value, to communicate between modules (reactiveValue)}

\item{m}{Shiny reactive value, to communicate between modules (reactiveValue)}

\item{session}{Shiny session variable}

\item{input}{Shiny input variable}

\item{output}{Shiny output variable}

\item{ns}{Shiny namespace}

\item{i18n}{Translator object from shiny.i18n library}

\item{delete_prefix}{Prefix used for reactive & input variables (character)}

\item{dialog_title}{Text for the title of the reactOutput (character)}

\item{dialog_subtext}{Text for the subText area of the reactOutput (character)}

\item{react_variable}{Name of r variable used to open or close reactOutput (character)}

\item{table}{Name of the table in the database where the data will be deleted (character)}

\item{r_table}{Name of the r variable containing the data to be deleted (charadter)}

\item{id_var_sql}{Name of the ID column in the database (character)}

\item{id_var_r}{Name of the ID column in r variable (character)}

\item{delete_message}{Message displayed when the deletion is confirmed (character)}

\item{reload_variable}{Name of r variable that triggers the reload of data (character)}

\item{information_variable}{Name of r variable containing the ID of deleted data (character)}

\item{translation}{TRUE if dialog_title & dialog_subtext have to be translated (logical)}

\item{app_folder}{App folder (character)}

\item{prefix}{A prefix used in Shiny modules (character)}

\item{r_message_bar}{If the MessageBar is displayed with a r variable or with show_message_bar function (TRUE for r variable)}
}
\description{
Create reactivity to delete a row in a datatable
}
\examples{
\dontrun{
widget_delete_prefix <- paste0(prefix, "_widget")
widget_dialog_title <- paste0(prefix, "_widgets_delete")
widget_dialog_subtext <- paste0(prefix, "_widgets_delete_subtext")
widget_react_variable <- "widget_delete_confirm"
widget_table <- paste0(prefix, "_widgets")
widget_id_var_sql <- "id"
widget_id_var_r <- paste0(prefix, "_selected_widget")
widget_delete_message <- paste0(prefix, "_widget_deleted")
widget_reload_variable <- paste0(prefix, "_load_ui")
widget_delete_variable <- paste0(widget_delete_prefix, "_open_dialog")
widget_information_variable <- paste0(prefix, "_widget_deleted")

delete_element(r = r, input = input, output = output, session = session, ns = ns, i18n = i18n,
  delete_prefix = widget_delete_prefix, dialog_title = widget_dialog_title, dialog_subtext = widget_dialog_subtext,
  react_variable = widget_react_variable, table = widget_table, id_var_sql = widget_id_var_sql, id_var_r = widget_id_var_r,
  delete_message = widget_delete_message, translation = TRUE, reload_variable = widget_reload_variable,
  information_variable = widget_information_variable)
}
}
