% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{delete_settings_datatable_row}
\alias{delete_settings_datatable_row}
\title{Delete a row in datatable}
\usage{
delete_settings_datatable_row(
  output,
  id = character(),
  r = shiny::reactiveValues(),
  ns = character(),
  i18n = character(),
  link_id = integer(),
  category = character(),
  row_deleted = integer(),
  table = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{r Shiny reactive value used to communicate between tabs}

\item{ns}{Shiny namespace}

\item{row_deleted}{ID of row to delete (integer)}

\item{table}{Name of the table used (character)}

\item{language}{Language used (character)}
}
\description{
Delete a row in datatable
}
\examples{
\dontrun{
delete_settings_datatable_row(output = output, r = r, ns = ns, language = "EN", row_deleted = 13, table = "datasets")
}
}
