% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_plugins.R
\name{display_plugin_error}
\alias{display_plugin_error}
\title{Display plugin error message}
\usage{
display_plugin_error(output, language, message)
}
\arguments{
\item{message}{Error message to display (character)}
}
\description{
Display plugin error message
}
