% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings.R
\name{execute_settings_code}
\alias{execute_settings_code}
\title{Execute / test code after edition}
\usage{
execute_settings_code(
  output,
  r = shiny::reactiveValues(),
  edited_code = character()
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{edited_code}{New code, after editing it (character)}
}
\description{
Execute / test code after edition
}
