% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{forbidden_card}
\alias{forbidden_card}
\title{Forbidden card}
\usage{
forbidden_card(
  ns = shiny::NS(),
  name = character(),
  language = "EN",
  words = tibble::tibble()
)
}
\description{
Forbidden card
}
