% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{format_datetime}
\alias{format_datetime}
\title{Format datetime}
\usage{
format_datetime(datetime = character(), language = "en", sec = TRUE)
}
\arguments{
\item{datetime}{Datetime converted in character vector (character)}

\item{language}{Selected language, "en" or "fr" (character)}

\item{sec}{Should we display seconds or not (logical)}
}
\description{
Format datetime depending on the selected language
}
\details{
English format of datetime is like "2023-05-20 16:56:22"\cr
French format is like "20-05-2023 16:56:22".\cr
Datetime has to be given in the default format ("\%Y-\%m-\%d \%H:\%M:\%S", like "2023-05-20 16:56:22").
}
\examples{
\dontrun{
  format_datetime(datetime = "2023-05-20 16:56:22", language = "fr", sec = FALSE)
}
}
