% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_authorized_data}
\alias{get_authorized_data}
\title{Get authorized data for a user}
\usage{
get_authorized_data(r, table, data = tibble::tibble())
}
\arguments{
\item{r}{Shiny reactive value, used to communicate between modules}

\item{table}{Name of the table in the database (character)}

\item{data}{If data is not r[[table]] (tibble / dataframe)}
}
\description{
Get authorized data for a user
}
\details{
Some data have a restricted access. This function filter the data the user has access to.
}
\examples{
\dontrun{
get_authorized_data(r = r, table = "plugins")
}
}
