% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_db}
\alias{get_db}
\title{Get database DBI object}
\usage{
get_db(
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  app_db_folder = character()
)
}
\arguments{
\item{r}{Shiny reactive value, used to communicate between modules}

\item{m}{Shiny reactive value, used to communicate between modules}

\item{app_db_folder}{Folder where the application database is saved (character)}
}
\description{
Get final database connection DBI object (local or remote).
}
\examples{
\dontrun{
get_db(r = r, m = m, app_db_folder = paste0(r$app_folder, "/app_database"))
}
}
