% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_local_db}
\alias{get_local_db}
\title{Connection to local database}
\usage{
get_local_db(app_db_folder = character())
}
\description{
Get a connection to local application database. If tables do not already exist, there are created
(with db_create_tables function). It uses RSQLite library.
It also adds distant database connection informations in the options table, if they do not already exist.
}
