% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_page_options}
\alias{get_page_options}
\title{Get options of a page}
\usage{
get_page_options(id = character())
}
\arguments{
\item{id}{ID of the tab / page}
}
\value{
A character vector with options
}
\description{
Get the options of a setting page (as datasets, studies...)
}
\examples{
get_page_options(id == "settings_datasets")
}
