% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{get_plural}
\alias{get_plural}
\title{Get plural form of a word}
\usage{
get_plural(word = character(), language = "en")
}
\arguments{
\item{word}{Original word, in its singular form (character)}

\item{language}{language used for conversion (character)}
}
\description{
Returns the plural form of a word
}
\examples{
get_plural(word = "study", language = "en")
}
