% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_remote_db}
\alias{get_remote_db}
\title{Connection to remote database}
\usage{
get_remote_db(
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  output,
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{local_db}{DBI db object of local database}

\item{language}{Language used to display messages (character)}
}
\description{
Get a connection to a remote database. If the remote connection fails, returns local DBI connection object.
}
