% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_singular}
\alias{get_singular}
\title{Get singular form of a word}
\usage{
get_singular(word = character(), language = "EN")
}
\arguments{
\item{word}{Original word, in its plural form (character)}

\item{language}{language used for conversion (character)}
}
\description{
Returns the singular form of a word
}
\examples{
get_singular(word = "studies", language = "EN")
}
