% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{get_thesaurus_item}
\alias{get_thesaurus_item}
\title{Get thesaurus item}
\usage{
get_thesaurus_item(
  output,
  r = shiny::reactiveValues(),
  thesaurus_name = character(),
  item_name = character(),
  item_id = integer(),
  method = character(),
  create = FALSE,
  item_unit = NA_character_,
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{output}

\item{r}{r}

\item{thesaurus_name}{thesaurus_name}

\item{item_name}{item_name}

\item{item_id}{item_id}

\item{method}{method}

\item{create}{create}

\item{item_unit}{item_unit}

\item{i18n}{i18n}

\item{ns}{ns}
}
\description{
Get a thesaurus item
}
