% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{get_thesaurus_items_levels}
\alias{get_thesaurus_items_levels}
\title{Get thesaurus items levels}
\usage{
get_thesaurus_items_levels(
  data = tibble::tibble(),
  r = shiny::reactiveValues()
)
}
\arguments{
\item{data}{datatable containing thesaurus items}

\item{r}{A shiny::reactiValues object, used to communicate between modules}
}
\description{
Get levels of thesaurus items to create thesaurus tree
}
