% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{get_thesaurus_items_paths}
\alias{get_thesaurus_items_paths}
\title{Get thesaurus items paths}
\usage{
get_thesaurus_items_paths(data = tibble::tibble(), r = shiny::reactiveValues())
}
\arguments{
\item{data}{datatable containing thesaurus items}

\item{r}{A shiny::reactiValues object, used to communicate between modules}
}
\description{
Get paths of thesaurus items to create thesaurus tree
}
