% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_dataset}
\alias{import_dataset}
\title{Import a dataset}
\usage{
import_dataset(
  output,
  ns = character(),
  i18n = character(),
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  dataset_id = integer(),
  data = tibble::tibble(),
  type = "",
  omop_version = "6.0",
  save_as_csv = TRUE,
  rewrite = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{ns}{Shiny namespace}

\item{i18n}{shiny.i18n object for translations}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{d}{A shiny::reactiveValues object, used to communicate between modules}

\item{dataset_id}{ID of the dataset, used to create directory in data folder (integer)}

\item{data}{Data variable (data.frame or tibble)}

\item{type}{Name of the OMOP table to import (character)}

\item{omop_version}{OMOP versions of the imported data, accepts "5.3", "5.4" or "6.0" (character)}

\item{save_as_csv}{Save or not the data to CSV file (logical)}

\item{rewrite}{If save_as_csv is TRUE, rewrite or not existing CSV file (logical)}

\item{quiet}{Should message bars be displayed (logical)}
}
\description{
Import a dataset in the application, with OMOP CDM

Load +/- save data when a dataset is selected by the user.
}
\details{
The function is used in a dataset code and is launched each time a user selects a dataset. \cr\cr
You can choose to \strong{load data each time} the function is used with save_as_csv set to FALSE (eg when dataset is small and the
data is loaded from a data warehouse) or you can \strong{save the data in a CSV file} with save_as_csv set to TRUE. \cr\cr
The data you import must respect the format of \href{https://ohdsi.github.io/CommonDataModel/cdm60.html}{\strong{OMOP common data model}}.\cr\cr
See help pages in the app for more informations.
}
\examples{
\dontrun{
persons <- tibble::tibble(person_id = 1L, gender_concept_id = 8532L, year_of_birth = NA_integer_, month_of_birth = NA_integer_,
  day_of_birth = NA_integer_, birth_datetime = lubridate::ymd_hms("1990-01-03 00:00:00"), race_concept_id = NA_integer_,
  ethnicity_concept_id = NA_integer_, location_id = NA_integer_, provider_id = NA_integer_, care_site_id = NA_integer_,
  person_source_value = NA_character_, gender_source_value = "F", race_source_value = NA_character_, ethnicity_source_value = NA_character_)
    
import_dataset(output = output, ns = ns, i18n = i18n, r = r, d = d, dataset_id = 5, data = persons, type = "persons", 
  omop_versions = "5.4", save_as_csv = FALSE, rewrite = FALSE, quiet = TRUE)
}
}
