% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_thesaurus}
\alias{import_thesaurus}
\title{Import a thesaurus}
\usage{
import_thesaurus(
  output,
  r = shiny::reactiveValues(),
  thesaurus_id = integer(),
  thesaurus = tibble::tibble(),
  language = "EN"
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{thesaurus_id}{ID of the thesaurus, used to update thesaurus table (integer)}

\item{thesaurus}{thesaurus data variable (data.frame or tibble)}

\item{language}{language used for error / warning messages (character, default = "EN")}
}
\description{
Add new thesaurus items to database
}
\details{
The function is used in a thesaurus code, it is launched only when you click on "Run code" on the thesaurus page.\cr\cr
Columns needed  :\cr\cr
\itemize{
\item{item_id = integer}
\item{name = character}
\item{display_name = character}
\item{category = character}
\item{unit = character}
}
}
\examples{
\dontrun{
thesaurus <- tibble::tribble(~item_id, ~name, ~display_name, ~category, ~unit,
  44543L, "Heart rate", "HR", "Vitals", "bpm",
  46531L, "Respiratory rate", "RR", "Vitals", "cpm")
  
import_thesaurus(output = output, r = r, thesaurus_id = 5, thesaurus = thesaurus, language = language)
}
}
