% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_vocabulary_table}
\alias{import_vocabulary_table}
\title{Import a vocabulary table}
\usage{
import_vocabulary_table(
  output,
  ns = character(),
  i18n = character(),
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  table_name = character(),
  data = tibble::tibble(),
  vocabulary_id = character(),
  messages_bars = FALSE
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{ns}{Shiny namespace}

\item{i18n}{shiny.i18n object for translations}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{m}{A shiny::reactiveValues object, used to communicate between modules}

\item{table_name}{Name of the vocabulary table we import (concept, concept_relationship or other) (character)}

\item{data}{A tibble containing the data}

\item{vocabulary_id}{ID of the vocabulary, for example LOINC (character)}

\item{message_bars}{Render messages in message bars or not (logical)}
}
\description{
Import an OMOP vocabulary and save it in app database
}
\details{
The function is used in a vocabulary code, it is launched only when you click on "Run code" on the vocabulary page.\cr\cr
See \href{https://ohdsi.github.io/CommonDataModel/cdm60.html}{\strong{OMOP common data model}} for more information.
}
\examples{
\dontrun{
concept <- tibble::tibble(concept_id = 3027018, concept_name = "Heart rate", domain_id = "Measurement",
  concept_class_id = "Clinical Observation", standard_concept = "S", concept_code = "8867-4",
  valid_start_date = "1970-01-01", valid_end_date = "2099-12-31", invalid_reason = NA_character_)
  
import_vocabulary_table(output = output, ns = ns, i18n = i18n, r = r, m = m,
  table_name = "concept", data = concept, vocabulary_id = "LOINC", message_bars = FALSE)
}
}
