% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_default_data.R
\name{insert_default_data}
\alias{insert_default_data}
\title{Insert default values in database}
\usage{
insert_default_data(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  i18n = character(),
  has_internet = FALSE,
  db_col_types = character(),
  users_accesses_toggles_options = tibble::tibble()
)
}
\arguments{
\item{r}{shiny reactive value, used to communicate between modules}
}
\description{
Insert default values in database
}
