% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_default_values.R
\name{insert_default_values}
\alias{insert_default_values}
\title{Insert default values in database}
\usage{
insert_default_values(output, r)
}
\arguments{
\item{r}{shiny reactive value, used to communicate between modules}
}
\description{
Insert default values in database
}
