% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkr.R
\name{linkr}
\alias{linkr}
\title{Run the Shiny Application}
\usage{
linkr(
  language = "en",
  app_folder = character(),
  perf_monitoring = FALSE,
  debug = FALSE,
  local = FALSE
)
}
\arguments{
\item{language}{Default language to use in the App (character)}

\item{app_folder}{Location of the application folder (character).}

\item{perf_monitoring}{Monitor app performances (boolean)}

\item{debug}{Debug mode : steps and errors will by displayed in the console (boolean)}

\item{local}{Run the app in local mode, do not load files on the internet (boolean)}
}
\description{
Runs the LinkR Shiny Application.\cr
Use language argument to choose language to use ("en" or "fr" available).\cr
Use app_folder argument to choose a folder where the files of the application will be saved. By default, 
a folder will be created depending of the value 'path.expand("~")' ('path.expand("~")/linkr').\cr
}
\examples{
\dontrun{
linkr(language = "en", perf_monitoring = FALSE, debug = FALSE, local = FALSE)
}
}
