% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{load_database}
\alias{load_database}
\title{Load database}
\usage{
load_database(r = shiny::reactiveValues(), language = "EN")
}
\arguments{
\item{r}{Shiny r reactive value, used to communicate between modules}
}
\description{
Load database
}
