% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_card}
\alias{make_card}
\title{Make a shiny.fluent card}
\usage{
make_card(title = character(), content = character(), size = 12, style = "")
}
\arguments{
\item{title}{Title of the card (character)}

\item{content}{Content of the card (character)}

\item{size}{Size of a card (integer)}

\item{style}{CSS code to custom the card (character)}
}
\value{
Shiny UI elements / HTML code
}
\description{
Code available in shiny.fluent github pages (Articles). Creates a shiny.fluent card.
}
\examples{
make_card(title = "Introduction", content = "This is the text of my introduction card", size = 12)
}
