% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_choicegroup}
\alias{make_choicegroup}
\title{Make a shiny.fluent choiceGroup}
\usage{
make_choicegroup(
  language = "EN",
  ns = character(),
  label = character(),
  id = NA_character_,
  options = list(),
  value = character(),
  inline = FALSE
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{label}{Label used for the input (character)}

\item{id}{ID used for the input (character)}

\item{options}{Options available for the choiceGroup (list)}

\item{value}{Value of the choiceGroup (character)}

\item{inline}{Is the choiceGroup displayed inline ? (logical)}
}
\description{
Make a shiny.fluent choiceGroup
}
