% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_combobox}
\alias{make_combobox}
\title{Make a shiny.fluent combo box}
\usage{
make_combobox(
  language = "EN",
  ns = shiny::NS(),
  label = character(),
  options = list(),
  multiSelect = TRUE,
  allowFreeform = FALSE,
  autoComplete = "on",
  id = NA_character_,
  value = NULL,
  width = NULL
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{label}{Label used for the input (character)}

\item{options}{Options available for the dropdown (list)}

\item{multiSelect}{Is multiselection of options is possible ? (logical)}

\item{id}{ID used for the input (character)}

\item{value}{Value of the toggle : TRUE or FALSE (logical)}

\item{width}{Width of the dropdown, CSS code so "300px" or "100\%" are accepted}

\item{allowFreeForm}{Allows user to enter free text, not provided by options (logical)}
}
\description{
Make a shiny.fluent combo box
}
