% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_dropdown}
\alias{make_dropdown}
\title{Make dropdown}
\usage{
make_dropdown(
  i18n = character(),
  ns = character(),
  id = NA_character_,
  label = character(),
  options = list(),
  value = NULL,
  multiSelect = FALSE,
  width = NULL,
  disabled = FALSE
)
}
\arguments{
\item{i18n}{Translator object from shiny.i18n library}

\item{ns}{Shiny namespace}

\item{id}{ID used for the input (character)}

\item{label}{Label used for the input (character)}

\item{options}{Options available for the dropdown (list)}

\item{value}{Value of the dropdown (character)}

\item{multiSelect}{Is multiselection of options is possible ? (logical)}

\item{width}{Width of the dropdown, CSS code so "300px" or "100\%" are accepted (character)}

\item{disabled}{Is the textfield is disabled (character)}
}
\description{
Creates a shiny.fluent dropdown
}
\examples{
\dontrun{
options <- list(
  list(key = "my_key1", text = "my_text1"),
  list(key = "my_key2", text = "my_text2")
)
make_dropdown(i18n = i18n, ns = NS("settings_datasets"), id = "my_dropdown", label = "my_dropdown",
  options = options, value = "my_key1", multiSelect = FALSE, width = "100\%", disabled = FALSE)
}
}
