% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_page}
\alias{make_page}
\title{Make a page}
\usage{
make_page(title = character(), subtitle = character(), contents = character())
}
\arguments{
\item{title}{Title of the page (character)}

\item{subtitle}{Subtitle of the page (character)}

\item{contents}{Contents of the page (character)}
}
\value{
Shiny UI elements / HTML code
}
\description{
Code available in shiny.fluent github pages (Articles). Creates a shiny.fluent page.
}
\examples{
make_page(title = "My page title", subtitle = "My page subtitle", contents = "shiny::div('My page content')")
}
