% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_people_picker}
\alias{make_people_picker}
\title{Make a shiny.fluent people picker}
\usage{
make_people_picker(
  language,
  ns,
  label,
  options,
  value = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL
)
}
\description{
Make a shiny.fluent people picker
}
