% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_textfield}
\alias{make_textfield}
\title{Make a shiny.fluent textfield}
\usage{
make_textfield(
  language,
  ns,
  label,
  id = NULL,
  value = NULL,
  type = NULL,
  canRevealPassword = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  margin_right = NULL
)
}
\description{
Make a shiny.fluent textfield
}
