% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_textfield_new}
\alias{make_textfield_new}
\title{Make a shiny.fluent textfield}
\usage{
make_textfield_new(
  i18n = R6::R6Class(),
  ns = shiny::NS(),
  label = character(),
  id = NA_character_,
  value = NULL,
  type = NULL,
  canRevealPassword = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  margin_right = NULL
)
}
\value{
Shiny UI elements / HTML code
}
\description{
Make a shiny.fluent textfield
}
\examples{

}
