% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_toggle}
\alias{make_toggle}
\title{Make a shiny.fluent toggle}
\usage{
make_toggle(
  language = "EN",
  ns = shiny::NS(),
  label = character(),
  id = NULL,
  value = FALSE,
  inline = FALSE,
  translate = TRUE,
  words = tibble::tibble()
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{label}{Label used for the input (character)}

\item{id}{ID used for the input (character)}

\item{value}{Value of the toggle : TRUE or FALSE (logical)}

\item{inline}{Should the toggle displayed inline (logical)}
}
\value{
HTML / Shiny UI code
}
\description{
Make a shiny.fluent toggle
}
\examples{
\dontrun{
make_toggle(language = "EN", ns = ns, label = "My toggle", id = "my_toggle", value = TRUE, inline = FALSE)
}
}
