% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{monitor_perf}
\alias{monitor_perf}
\title{Monitor performances}
\usage{
monitor_perf(r = shiny::reactiveValues(), action = "stop", task = character())
}
\arguments{
\item{r}{Shiny reactive value, used to communicate between modules (reactiveValue)}

\item{action}{If the function is called at the beginning of the end of a task, takes "start" or "stop" values (character)}

\item{task}{Name of the task (character)}
}
\description{
Monitor performances of the ShinyApp, with calculating time each step takes
}
\examples{
\dontrun{
 monitor_perf(r = r, action = "start", task = "Load database")
}
}
