% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{prepare_data_datatable_new}
\alias{prepare_data_datatable_new}
\title{Reload cache}
\usage{
prepare_data_datatable_new(
  output,
  r = shiny::reactiveValues(),
  ns = shiny::NS(),
  i18n = R6::R6Class(),
  id = character(),
  table = character(),
  dropdowns = character(),
  dropdowns_multiselect = character(),
  dropdowns_null_value = character(),
  factorize_cols = character(),
  action_buttons = character(),
  data_input = tibble::tibble(),
  data_output = tibble::tibble(),
  words = tibble::tibble()
)
}
\arguments{
\item{r}{Shiny r reactive value}

\item{table}{Database table name}

\item{data}{Data}
}
\description{
Reload cache
}
