% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{reload_cache_for_settings}
\alias{reload_cache_for_settings}
\title{Reload cache}
\usage{
reload_cache_for_settings(
  r = shiny::reactiveValues(),
  table = character(),
  data = tibble::tibble()
)
}
\arguments{
\item{r}{Shiny reactive value}

\item{table}{Database table name}

\item{data}{Data}
}
\description{
Reload cache
}
