% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{remove_patients_from_subset}
\alias{remove_patients_from_subset}
\title{Remove patients from a subset}
\usage{
remove_patients_from_subset(
  output,
  r = shiny::reactiveValues(),
  patients = tibble::tibble(),
  subset_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between tabs}

\item{patients}{data variable containing patients (data.frame / tibble)}

\item{subset_id}{ID of subset (integer)}

\item{i18n}{shiny.i18n object for translations}

\item{ns}{shiny namespace object, used to render messages on message bars}
}
\description{
Remove patients from a subset
}
\examples{
\dontrun{
patients <- tibble::tribble(~patient_id, 123L, 456L, 789L)
remove_patients_from_subset(output = output, r = r, patients = patients, subset_id = 3, i18n = i18n, ns = ns)
}
}
