% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{remove_persons_from_subset}
\alias{remove_persons_from_subset}
\title{Remove patients from a subset}
\usage{
remove_persons_from_subset(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  persons = tibble::tibble(),
  subset_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{m}{A shiny::reactiveValues object, used to communicate between modules}

\item{persons}{data variable containing patients / persons (data.frame / tibble)}

\item{subset_id}{ID of subset (integer)}

\item{i18n}{Translator object from shiny.i18n library}

\item{ns}{Shiny namespace}
}
\description{
Remove patients from a subset
}
\examples{
\dontrun{
persons <- tibble::tribble(~patient_id, 123L, 456L, 789L)
remove_persons_from_subset(output = output, r = r, m = m, persons = persons, subset_id = 3, i18n = i18n, ns = ns)
}
}
