% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{render_datatable}
\alias{render_datatable}
\title{Render a DT datatable}
\usage{
render_datatable(
  output,
  r = shiny::reactiveValues(),
  ns = shiny::NS(),
  language = "EN",
  data = tibble::tibble(),
  output_name = character(),
  col_names = character(),
  datatable_dom = "<'datatable_length'l><'top'ft><'bottom'p>",
  page_length = 10,
  start = 1,
  editable_cols = character(),
  sortable_cols = character(),
  centered_cols = character(),
  searchable_cols = character(),
  filter = FALSE,
  factorize_cols = character(),
  column_widths = character()
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{ns}{Shiny namespace}

\item{language}{Language used (charater)}

\item{data}{data used in the datatable (tibble or dataframe)}

\item{output_name}{Name of the datatable output}

\item{col_names}{A character vector containing colnames, already translated (character)}

\item{datatable_dom}{Character containing DOM code for the datatable (character)}

\item{page_length}{Page length of the datatable, default to 10 rows (integer)}

\item{start}{Which page display (used when we save datatable state), default to 1 (integer)}

\item{editable_cols}{Which cols are editable (character vector)}

\item{sortable_cols}{Which cols are sortable (character vector)}

\item{centered_cols}{Which cols are centered (character vector)}

\item{searchable_cols}{If filter is TRUE, choose which columns are searchable (character)}

\item{filter}{If TRUE, we can filter we search box each column (logical)}

\item{factorize_cols}{Which columns are factorized (to be filtered with a dropdown) (character)}

\item{column_widths}{Columns widths (named character vector)}
}
\description{
Renders a datatable (from library DT)
}
