% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_code_card}
\alias{render_settings_code_card}
\title{Render UI of edit_code card}
\usage{
render_settings_code_card(
  ns = shiny::NS(),
  id = character(),
  title = character(),
  code = character(),
  link_id = integer(),
  language = "EN"
)
}
\arguments{
\item{ns}{Shiny namespace}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{title}{Title of the card (character)}

\item{code}{Code to show in ShinyAce editor (character)}

\item{link_id}{ID allows to link with code table (integer)}

\item{language}{Language used (character)}
}
\description{
Render UI of edit_code card
}
\examples{
\dontrun{
render_settings_code_card(ns = NS("settings_datamarts"), id = "settings_datamarts", title = "edit_datamart_code",
  code = "Enter your code here", link_id = 3, language = "EN")
}
}
