% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{render_settings_datatable}
\alias{render_settings_datatable}
\title{Render management datatable}
\usage{
render_settings_datatable(
  output,
  r = shiny::reactiveValues(),
  id = character(),
  col_names = character(),
  table = character(),
  dropdowns = character(),
  action_buttons = character(),
  datatable_dom = "<'datatable_length'l><'top'ft><'bottom'p>",
  page_length = 10,
  start = 1,
  editable_cols = integer(),
  column_defs = character()
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{col_names}{A character vector containing colnames, already translated (character)}

\item{table}{Name of a table of database (character)}

\item{dropdowns}{Character vector with names of the dropdowns available in the datatable (character)}

\item{action_buttons}{Character vector with action_buttons needed (character)}

\item{datatable_dom}{Character containing DOM code for the datatable (character)}

\item{page_length}{Page length of the datatable, default to 10 rows (integer)}

\item{start}{Which page display (used when we save datatable state), default to 1 (integer)}

\item{editable_cols}{Which cols are editable (numeric vector)}

\item{column_defs}{DataTable columnDefs parameters (character)}
}
\description{
Render management datatable
}
