% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_datatable_card}
\alias{render_settings_datatable_card}
\title{Render UI of settings datatable card}
\usage{
render_settings_datatable_card(
  language = "EN",
  ns = shiny::NS(),
  output_id = "management_datatable",
  title = character()
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{output_id}{ID of DTOutput, allows to have multiple management_datatable in one module, default = "management_datatable" (character)}

\item{title}{Title used to create the card, it will be translated with translate function (character)}
}
\description{
Render UI of settings datatable card
}
\examples{
\dontrun{
render_settings_datatable_card(language = "EN", ns = ns, output_id = "management_datatable", title = "datamarts_management")
}
}
