% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_default_elements}
\alias{render_settings_default_elements}
\title{Render UI of settings default elements}
\usage{
render_settings_default_elements(ns = shiny::NS())
}
\arguments{
\item{ns}{Shiny namespace}
}
\value{
Shiny UI elements / HTML code
}
\description{
Set default UI elements on top of the page : message_bar outputs, react output to confirm delete of a table element
}
\examples{
render_settings_default_elements(ns = NS("settings_datamart"))
}
