% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{render_settings_delete_react}
\alias{render_settings_delete_react}
\title{Render delete react}
\usage{
render_settings_delete_react(
  r = shiny::reactiveValues(),
  ns = character(),
  table = character(),
  i18n = character()
)
}
\arguments{
\item{r}{Shiny r reactive value to communicate between modules}

\item{ns}{Shiny namespace}

\item{table}{Name of the table used (character)}

\item{language}{Language used (character)}
}
\description{
Render delete react
}
\examples{
\dontrun{
render_settings_delete_react(r = r, table = "datamarts")
}
}
