% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_options_card}
\alias{render_settings_options_card}
\title{Render UI of options card}
\usage{
render_settings_options_card(
  ns = shiny::NS(),
  r = r,
  id = character(),
  title = character(),
  code = character(),
  category = character(),
  link_id = integer(),
  language = "EN",
  words = tibble::tibble()
)
}
\arguments{
\item{ns}{Shiny namespace}

\item{r}{Shiny r reactive value}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{title}{Title of the card (character)}

\item{code}{Code to show in ShinyAce editor (character)}

\item{category}{Category allowing to link with code table (character)}

\item{link_id}{ID allowing to link with code table (integer)}

\item{language}{Language used (character)}
}
\description{
Render UI of options card
}
