% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_toggle_card}
\alias{render_settings_toggle_card}
\title{Render UI of settings toggle card}
\usage{
render_settings_toggle_card(
  language = "EN",
  ns = shiny::NS(),
  cards = list(),
  activated = "",
  translate = TRUE,
  words = tibble::tibble()
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{cards}{A list containing distinct cards (list)}

\item{activated}{Which toggles are set to ON (character)}
}
\value{
Shiny UI elements / HTML code
}
\description{
At the top of main page, there's a card with toggle buttons, it allows to show or hide distinct cards of the page.
}
\examples{
cards <- list(
  list(key = "creation_card", label = "create_datamart"),
  list(key = "datatable_card", label = "datamarts_management"),
  list(key = "edit_code_card", label = "edit_datamart_code"))
settings_toggle_card(ns = NS("settings_datamart"), cards = cards, activated = c("creation_card", "datatable_card"), language = "EN")
}
