% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{run_datamart_code}
\alias{run_datamart_code}
\title{Run datamart code}
\usage{
run_datamart_code(
  output,
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  datamart_id = integer(),
  i18n = character(),
  quiet = TRUE
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between modules}

\item{d}{A shiny::reactiValues object, used to communicate between modules. Contains data loaded from datamart code (d$patients, d$labs_vitals...).}

\item{datamart_id}{ID of datamart containing the code (integer)}

\item{i18n}{shiny.i18n object for translations}
}
\description{
Runs datamart code
}
\details{
The code to load the datamart is loaded from the application database, in code table, from datamart_id.
}
\examples{
\dontrun{
run_datamart_code(output = output, r = r, d = d, i18n = i18n, datamart_id = 3)
}
}
