% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{run_dataset_code}
\alias{run_dataset_code}
\title{Run dataset code}
\usage{
run_dataset_code(
  output,
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  dataset_id = integer(),
  i18n = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{d}{A shiny::reactiveValues object, used to communicate between modules. Contains data loaded from dataset code (d$patients, d$labs_vitals...).}

\item{dataset_id}{ID of the dataset we want to load (integer)}

\item{i18n}{Translator object from shiny.i18n library}
}
\description{
Runs the code of a dataset
}
\examples{
\dontrun{
run_dataset_code(output = output, r = r, d = d, dataset_id = 3, i18n = i18n)
}
}
