% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{save_settings_datatable_updates}
\alias{save_settings_datatable_updates}
\title{Save changes in datatable}
\usage{
save_settings_datatable_updates(
  output,
  r = shiny::reactiveValues(),
  ns = shiny::NS(),
  table = character(),
  duplicates_allowed = FALSE,
  language = "EN"
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{Shiny r reactive value to communicate between modules}

\item{ns}{Shiny namespace}

\item{table}{Name of the table used (character)}

\item{duplicates_allowed}{Are duplicates in the name column allowed (logical)}

\item{language}{Language used (character)}
}
\description{
Save changes in datatable
}
\examples{
\dontrun{
save_settings_datatable_updates(output = output, r = r, ns = ns, table = "datamarts", language = "EN")
}
}
