% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{save_settings_options}
\alias{save_settings_options}
\title{Save options}
\usage{
save_settings_options(
  output,
  r = shiny::reactiveValues(),
  id = character(),
  category = character(),
  code_id_input = integer(),
  data = data,
  i18n = character(),
  page_options = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{Shiny reactive value, used to communicate between modules (reactiveValue)}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{category}{Category column in code table, eg : "dataset", "plugin" (character)}

\item{code_id_input}{Input of the actionButton containing ID of current row, in datatable, format = "edit_code_[ID]" (character)}

\item{data}{New data to store in options table (list)}

\item{i18n}{Translator object from shiny.i18n library}

\item{page_options}{Which options need to be saved (character)}
}
\description{
Save options in settings pages
}
\examples{
\dontrun{
data <- list()
data$show_only_aggregated_data <- TRUE
data$users_allowed_read <- c(1, 3, 4)
save_settings_options(output = output, r = r, id = "settings_dataset", category = "dataset", code_id_input = "edit_code_3",
  data = data, i18n = i18n, page_options = c("show_only_aggregated_data", "users_allowed_read"))
}
}
