% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings.R
\name{settings_edit_code_card}
\alias{settings_edit_code_card}
\title{Render UI of edit_code card}
\usage{
settings_edit_code_card(
  ns = shiny::NS(),
  id = character(),
  title = character(),
  code = character(),
  link_id = integer(),
  language = "EN"
)
}
\arguments{
\item{ns}{Shiny namespace}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{title}{Title of the card (character)}

\item{code}{Code to show in ShinyAce editor (character)}

\item{link_id}{ID allows to link with code table (integer)}

\item{language}{Language used (character)}
}
\description{
Render UI of edit_code card
}
