% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings.R
\name{settings_edit_code_save}
\alias{settings_edit_code_save}
\title{Save code edition}
\usage{
settings_edit_code_save(
  output,
  r = shiny::reactiveValues(),
  id = character(),
  code_id_input = integer(),
  code_edited = character(),
  language = "EN"
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{code_id_input}{Input of the actionButton containing ID of current row, in datatable, format = "edit_code_[ID]" (character)}

\item{code_edited}{New code, after editing it (character)}

\item{language}{Language used}
}
\description{
Save code edition
}
