% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{show_hide_cards}
\alias{show_hide_cards}
\title{Show or hide cards}
\usage{
show_hide_cards(
  r = shiny::reactiveValues(),
  session,
  input,
  table = character(),
  id = character(),
  toggles = character()
)
}
\description{
Show or hide cards
}
