% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{show_message_bar}
\alias{show_message_bar}
\title{Display a message bar}
\usage{
show_message_bar(
  output,
  message = character(),
  type = "severeWarning",
  i18n = character(),
  time = 7000,
  ns = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{message}{Message that will be displayed, after translation (character)}

\item{type}{Type of message bar displayed (reference to Microsoft MessageBarType num) (character)}

\item{i18n}{Translator object from shiny.i18n library}

\item{time}{Time the message bar will by displayed, in ms (integer)}

\item{ns}{Shiny namespace}
}
\description{
Displays a shiny.fluent message bar on the top of the page
}
\details{
The different possible types are : c("info", "error", "blocked", "severeWarning", "success", "warning")
}
\examples{
\dontrun{
message_bar(output = output, message = "name_already_used", type = "severeWarning", i18n = i18n, time = 5000, ns = ns)
}
}
