% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{show_or_hide_cards}
\alias{show_or_hide_cards}
\title{Show or hide cards}
\usage{
show_or_hide_cards(
  r = shiny::reactiveValues(),
  session,
  input,
  table = character(),
  id = character(),
  cards = character()
)
}
\arguments{
\item{r}{Shiny reactive value, used to communicate between modules (reactiveValue)}

\item{session}{Shiny session variable}

\item{input}{Shiny input variable}

\item{table}{Name of the database table (character)}

\item{id}{Name of the ID of the page (character)}

\item{cards}{Character vector containing names of current page cards (character)}
}
\description{
Show or hide cards depending of selected shiny.fluent::PivotItem
}
\examples{
\dontrun{
cards <- c("subsets_datatable_card", "subsets_persons_card", "subsets_edit_code_card")
show_or_hide_cards(r = r, input = input, session = session, id = id, cards = cards)
}
}
