% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{test_distant_db}
\alias{test_distant_db}
\title{Test connection to distant database}
\usage{
test_distant_db(local_db)
}
\arguments{
\item{local_db}{DBI db object of local database}
}
\description{
Test the connection to a distant database. If gets the distant db informations in local database
and then tries to connect to this distant database.
}
