% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_translate.R
\name{translate}
\alias{translate}
\title{Translates a word}
\usage{
translate(language, reference_word)
}
\arguments{
\item{language}{A two-letters or three-letters character indicating the requested language}

\item{words}{A tibble containing the translations}
}
\value{
Returns the translated word (character)
}
\description{
Translates a reference word to the requested language
}
\examples{
words <- dplyr::tribble(~language, ~reference_word, ~translated_word,
"FR", "Settings", "Parametres",
"IT", "Settings", "Impostazioni")

translate("ESP", words, "Settings")
translate("FR", words, "Settings")
}
