% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{update_r}
\alias{update_r}
\title{Update r variable}
\usage{
update_r(
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  table = character(),
  i18n = character()
)
}
\arguments{
\item{r}{A shiny reactiveValue, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{m}{A shiny reactiveValue, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{table}{Database table name (character)}

\item{i18n}{Translator object from shiny.i18n library}
}
\description{
Update r reactive variable, requesting the corresponding table in the database
}
\examples{
\dontrun{
update_r(r = r, table = "subsets", i18n = i18n)
}
}
