% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{update_settings_datatable}
\alias{update_settings_datatable}
\title{Update datatable}
\usage{
update_settings_datatable(
  input,
  tab_id = character(),
  r = shiny::reactiveValues(),
  ns = character(),
  table = character(),
  dropdowns = character(),
  i18n = character()
)
}
\arguments{
\item{input}{Shiny input variable}

\item{r}{Shiny reactive value, used to communicate between modules (reactiveValue)}

\item{ns}{Shiny namespace}

\item{table}{Name of the database table used (character)}

\item{dropdowns}{Dropdowns displayed on datatable (character)}

\item{i18n}{Translator object from shiny.i18n library}
}
\description{
Update data in r variable when an update is made in the datatable
}
\examples{
\dontrun{
update_settings_datatable(r = r, ns = ns, table = "datasets", dropdowns = "data_source", i18n = i18n)
}
}
