% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_patients_to_subset}
\alias{add_patients_to_subset}
\title{Add patients to a subset}
\usage{
add_patients_to_subset(
  output,
  r = shiny::reactiveValues(),
  patients = tibble::tibble(),
  subset_id = integer(),
  erase = FALSE,
  language = "EN"
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{patients}{data variable containing patients (data.frame / tibble)}

\item{subset_id}{ID of subset (integer)}

\item{erase}{if TRUE, erase patients contained in selected subset (logical)}

\item{language}{language used for error / warning messages (character, default = "EN")}
}
\description{
Add patients to a subset
}
\details{
...
}
\examples{
\dontrun{
patients <- tibble::tribble(~patient_id, 123L, 456L, 789L)
subset_add_patients(output = output, r = r, patients = patients, subset_id = 3, erase = FALSE, language = "EN")
}
}
