% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_persons_to_subset}
\alias{add_persons_to_subset}
\title{Add patients to a subset}
\usage{
add_persons_to_subset(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  persons = tibble::tibble(),
  subset_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between modules}

\item{m}{A shiny::reactiValues object, used to communicate between modules}

\item{subset_id}{ID of the subset (integer)}

\item{i18n}{shiny.i18n object for translations}

\item{ns}{shiny namespace object, used to render messages on message bars}

\item{patients}{data variable containing patients (data.frame / tibble)}
}
\description{
Add patients to a subset, only if not already in the subset
}
\examples{
\dontrun{
patients <- tibble::tribble(~patient_id, 123L, 456L, 789L)
subset_add_patients(output = output, r = r, m = m, patients = patients, subset_id = 3, i18n = i18n, ns = ns)
}
}
