% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{add_settings_new_data}
\alias{add_settings_new_data}
\title{Add new settings data}
\usage{
add_settings_new_data(
  session,
  output,
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  i18n = character(),
  id = character(),
  data = tibble::tibble(),
  table = character(),
  required_textfields = character(),
  req_unique_values = character(),
  required_dropdowns = "all",
  dropdowns = character(),
  r_message_bar = FALSE
)
}
\arguments{
\item{session}{Shiny session variable}

\item{output}{Shiny output variable}

\item{r}{Shiny reactive value, to communicate between modules (reactiveValue)}

\item{i18n}{Translator object from shiny.i18n library}

\item{id}{ID of current tab / page (character)}

\item{data}{A list with data to add (list)}

\item{table}{Name of the corresponding table in database (character)}

\item{required_textfields}{Which textfields are required (not empty) before inserting data in database ? (character)}

\item{req_unique_values}{Which fields require unique values before inserting data in database ? (character)}

\item{required_dropdowns}{Which dropdowns are required (not empty) before insert data in database ? Default to "all" (character)}

\item{dropdowns}{Tibble with the values of distinct dropdowns names (tibble)}

\item{r_message_bar}{If the MessageBar is displayed with a r variable or with show_message_bar function (TRUE for r variable)}
}
\description{
Add new settings data
}
\examples{
\dontrun{
data <- list()
data$name <- "New dataset"
data$description <- "Description of the dataset"
data$data_source <- 5
add_settings_new_data(output = output, r = r, i18n = i18n, id = "settings_datasets", data = data, table = "datasets", dropdowns = "data_source")
}
}
