% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{add_settings_new_data}
\alias{add_settings_new_data}
\title{Add new settings data}
\usage{
add_settings_new_data(
  session,
  output,
  r = shiny::reactiveValues(),
  language = "EN",
  id = character(),
  data = tibble::tibble(),
  table = character(),
  required_textfields = character(),
  req_unique_values = character(),
  required_dropdowns = "all",
  dropdowns = character()
)
}
\arguments{
\item{session}{Shiny session variable}

\item{output}{Shiny output variable}

\item{r}{Shiny r reactive value, to communicate between modules (reactiveValue)}

\item{language}{Language used (character)}

\item{id}{ID of current module / page (character)}

\item{data}{A list with data to add (list)}

\item{table}{Name of the corresponding table in database (character)}

\item{required_textfields}{Which textfields are required (not empty) before inserting data in database ? (character)}

\item{req_unique_values}{Which fields require unique values before inserting data in database ? (character)}

\item{required_dropdowns}{Which dropdowns are required (not empty) before insert data in database ? Default to "all" (character)}

\item{dropdowns}{Tibble with the values of distinct dropdowns names (tibble)}
}
\description{
Add new settings data
}
\examples{
\dontrun{
data <- list()
data$name <- "New datamart"
data$description <- "Description of the datamart"
data$data_source <- 5
add_settings_new_data(output = output, r = r, language = language, id = "settings_datamarts", data = data, dropdowns = "data_source")
}
}
