% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{convert_tibble_to_list}
\alias{convert_tibble_to_list}
\title{Convert a tibble to list}
\usage{
convert_tibble_to_list(
  data = tibble::tibble(),
  key_col = character(),
  text_col = character(),
  null_value = FALSE,
  language = "EN"
)
}
\arguments{
\item{data}{a tibble or a dataframe containing data}

\item{key_col}{name of the column containing the key (character)}

\item{text_col}{name of the column containing the text (character)}

\item{null_value}{add a null value (logical)}

\item{language}{language used (character)}
}
\value{
A list with this structure : list(list(key = "my_key1", text = "my_text1"), list(key = "my_key2", text = "my_text2"))
}
\description{
Convert a tibble to list
}
\examples{
data <- tibble::tribble(~key, ~text, "my_key1", "my_text1", "my_key2", "my_text2")
my_list <- convert_tibble_to_list(data = data, key_col = "key", text_col = "text", null_value = TRUE, language = "EN")
print(my_list)
}
