% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{create_datatable_cache}
\alias{create_datatable_cache}
\title{Create cache for datatable data}
\usage{
create_datatable_cache(
  r,
  table = character(),
  thesaurus_id = integer(),
  link_id = integer(),
  category = character()
)
}
\arguments{
\item{r}{Shiny r reactive value, to communicate between modules}

\item{table}{Table name of database (character)}

\item{thesaurus_id}{ID of thesaurus, which thesaurus items depend on (integer)}

\item{link_id}{Link ID to merge data from cache to original data (integer)}

\item{category}{Category of cache, depending of the page of Settings (character)}
}
\description{
Create cache for datatable data
}
