% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{create_scripts_thesaurus}
\alias{create_scripts_thesaurus}
\title{Create scripts thesaurus}
\usage{
create_scripts_thesaurus(
  output,
  r = shiny::reactiveValues(),
  data_source_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between tabs}

\item{data_source_id}{ID of the data source where the script is recorded}

\item{i18n}{shiny.i18n object for translations}

\item{ns}{shiny namespace object, used to render messages on message bars}
}
\description{
Create a thesaurus for the items created in the scripts
}
\details{
A script is a code executed each time we load a dataset.
In some scripts, we create new thesaurus items.
The scripts are attached to data sources. 
This function creates a new thesaurus, for a data source, named [data_source_name - scripts], which will include
all items created in the scripts depending of this data source.
}
\examples{
\dontrun{
# If my script depends of the data source with ID 13, named "My data source", this function will create a thesaurus
called "My data source - scripts".
create_scripts_thesaurus(output = output, r = r, data_source_id = 13, i18n = i18n, ns = ns)
}
}
