% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{db_create_table}
\alias{db_create_table}
\title{Create a new table}
\usage{
db_create_table(
  db,
  table_name = character(),
  dataframe = tibble::tibble(),
  dbms = character(),
  primary_key_col = character(),
  unique_cols = character(),
  not_null_cols = character(),
  text_cols = character()
)
}
\arguments{
\item{db}{Database connection object (DBI)}

\item{table_name}{Name of the new table (character)}

\item{dataframe}{Dataframe with table informations (columns names & type of value) (dataframe / tibble)}
}
\description{
Create a new table in the database
}
\examples{
\dontrun{
  db_create_table(db, "my_new_table", tibble::tibble(id = integer(), name = character()))
}
}
