% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{db_create_tables}
\alias{db_create_tables}
\title{Create tables}
\usage{
db_create_tables(db, type = character(), dbms = character())
}
\arguments{
\item{db}{A DBI database connection object}

\item{dbms}{Database management system used ("postgres" and "sqlite" supported) (character)}
}
\description{
Create all the tables used in the application.
}
\examples{
\dontrun{
db_create_tables(db = db, dbms = "postgres")
}
}
