% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{delete_element}
\alias{delete_element}
\title{Delete element}
\usage{
delete_element(
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  session,
  input,
  output,
  ns = character(),
  i18n = character(),
  delete_prefix = character(),
  dialog_title = character(),
  dialog_subtext = character(),
  react_variable = character(),
  table = character(),
  r_table = character(),
  id_var_sql = character(),
  id_var_r = character(),
  delete_message = character(),
  reload_variable = character(),
  information_variable = character(),
  translation = TRUE,
  app_folder = character(),
  prefix = character(),
  r_message_bar = FALSE
)
}
\description{
Delete element
}
