% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_col_names}
\alias{get_col_names}
\title{Get column names}
\usage{
get_col_names(table_name = character(), i18n = character())
}
\arguments{
\item{table_name}{Name of the table (character)}

\item{language}{Language used (charater)}
}
\description{
Get column names
}
\examples{
get_col_names(table_name = "datasets", language = "EN")
}
