% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_db}
\alias{get_db}
\title{Get database DBI object}
\usage{
get_db()
}
\description{
Get final database connection DBI object.
First, get local database connection. Second, if the choice recorded in local database is distant db, try distant db connection.
}
