% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_app_database.R
\name{get_last_row}
\alias{get_last_row}
\title{Get last row ID of a table}
\usage{
get_last_row(con, table)
}
\arguments{
\item{con}{DBI connection object to the database}

\item{table}{Name of the table}
}
\description{
Get last row ID of a table
}
