% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{get_thesaurus_item}
\alias{get_thesaurus_item}
\title{Get thesaurus item}
\usage{
get_thesaurus_item(
  output,
  r = shiny::reactiveValues(),
  thesaurus_name = character(),
  item_name = character(),
  item_id = integer(),
  method = character(),
  create = FALSE,
  item_unit = NA_character_,
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between modules}

\item{thesaurus_name}{Name of the thesaurus (character)}

\item{item_name}{Name of the item to get, if method = "item_name" (character)}

\item{item_id}{ID of the item to get, if method = "item_id" (integer)}

\item{method}{Method used to get the thesaurus item. Could be "item_name" or "item_id" (character).}

\item{create}{If the item is not found and create is TRUE, the item is created, with add_thesaurus_item function (logical).}

\item{item_unit}{Unit of the item, if it is not found and create is TRUE (character).}

\item{i18n}{shiny.i18n object for translations}

\item{ns}{shiny namespace object, used to render messages on message bars}
}
\description{
Get a thesaurus item, from the thesaurus name and ID or name of the item.
}
\examples{
\dontrun{
# Search item "Respiratory rate" from thesaurus "MIMIC-IV"
get_thesaurus_item(output = output, r = r, thesaurus_name = "MIMIC-IV", item_name = "Respiratory rate", method = "item_name", i18n = i18n, ns = ns)

# Search item 83748 from thesaurus "MIMIC-IV", and create it if not found
get_thesaurus_item(output = output, r = r, thesaurus_name = "MIMIC-IV", item_id = 83746, 
 item_name = "Heart rate", item_unit = "bpm", method = "item_id", i18n = i18n, ns = ns)
}
}
