% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_dropdown}
\alias{make_dropdown}
\title{Make a shiny.fluent dropdown}
\usage{
make_dropdown(
  language,
  ns,
  label,
  options = list(),
  multiSelect = FALSE,
  id = NULL,
  value = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  margin_right = NULL
)
}
\description{
Make a shiny.fluent dropdown
}
