% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_layout}
\alias{make_layout}
\title{Make layout}
\usage{
make_layout(
  language = "en",
  page = character(),
  i18n = character(),
  users_accesses_toggles_options = tibble::tibble()
)
}
\arguments{
\item{language}{Language used ("en" or "fr") (character)}

\item{page}{Name of the page (character)}

\item{i18n}{Translator object from shiny.i18n library}

\item{users_accesses_toggles_options}{A tibble containing users accesses, to add in database if no internet access (tibble)}
}
\description{
Creates a complete layout with header, sidenav, main & footer
}
\examples{
\dontrun{
make_layout(langauge = "fr", page = "my_subsets", i18n = i18n, users_accesses_toggles_options = users_accesses_toggles_options)
}
}
