% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_layout}
\alias{make_layout}
\title{Make a complete layout with header, sidenav, main & footer}
\usage{
make_layout(language = "EN", page = character(), words = tibble::tibble())
}
\description{
Make a complete layout with header, sidenav, main & footer
}
