% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_textfield}
\alias{make_textfield}
\title{Make a shiny.fluent textfield}
\usage{
make_textfield(
  language = "EN",
  ns = shiny::NS(),
  label = character(),
  id = NA_character_,
  value = NULL,
  type = NULL,
  canRevealPassword = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  margin_right = NULL,
  words = tibble::tibble()
)
}
\value{
Shiny UI elements / HTML code
}
\description{
Make a shiny.fluent textfield
}
