% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{make_toggle}
\alias{make_toggle}
\title{Make a shiny.fluent toggle}
\usage{
make_toggle(language, ns, label, id = NULL, value = FALSE, inline = FALSE)
}
\description{
Make a shiny.fluent toggle
}
