% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_code_card}
\alias{render_settings_code_card}
\title{Render UI of edit_code card}
\usage{
render_settings_code_card(
  ns = character(),
  r = shiny::reactiveValues(),
  id = character(),
  title = character(),
  code = list(),
  link_id = integer(),
  i18n = character()
)
}
\arguments{
\item{ns}{Shiny namespace}

\item{r}{r Shiny reactive value to communicate between modules}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{title}{Title of the card (character)}

\item{code}{Code to show in ShinyAce editor (list)}

\item{link_id}{ID allows to link with code table (integer)}

\item{i18n}{Translator object from shiny.i18n library}
}
\description{
Render UI of edit_code card
}
\examples{
\dontrun{
render_settings_code_card(ns = NS("settings_datasets"), r = r, id = "settings_datasets", title = "edit_dataset_code",
  code = "Enter your code here", link_id = 3, i18n = i18n)
}
}
