% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_datatable_card}
\alias{render_settings_datatable_card}
\title{Render UI of settings datatable card}
\usage{
render_settings_datatable_card(
  language = "EN",
  ns = shiny::NS(),
  id = character(),
  title = character()
)
}
\arguments{
\item{language}{Language used (character)}

\item{ns}{Shiny namespace}

\item{id}{ID of current module / page (character)}

\item{title}{Title used to create the card, it will be translated with translate function (character)}
}
\description{
Render UI of settings datatable card
}
\examples{
\dontrun{
render_settings_datatable_card(language = "EN", ns = ns, id = "settings_datamart", title = "datamarts_management")
}
}
