% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_ui.R
\name{render_settings_datatable_card}
\alias{render_settings_datatable_card}
\title{Render UI of settings datatable card}
\usage{
render_settings_datatable_card(
  i18n = character(),
  ns = character(),
  div_id = "datatable_card",
  output_id = "management_datatable",
  title = character(),
  inputs = character(),
  dropdown_multiselect = FALSE
)
}
\arguments{
\item{ns}{Shiny namespace}

\item{div_id}{ID of the div, to show or hide with toggles, default = "datatable_card" (character)}

\item{output_id}{ID of div & DTOutput, allows to have multiple management_datatable in one module, default = "management_datatable" (character)}

\item{title}{Title used to create the card, it will be translated with translate function (character)}

\item{language}{Language used (character)}
}
\description{
Render UI of settings datatable card
}
\examples{
\dontrun{
render_settings_datatable_card(language = "EN", ns = ns, output_id = "management_datatable", title = "datasets_management")
}
}
