% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_log.R
\name{report_bug}
\alias{report_bug}
\title{Report bug}
\usage{
report_bug(
  r = shiny::reactiveValues(),
  output,
  error_message = character(),
  error_name = character(),
  category = character(),
  error_report = character(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{r}{Shiny reactive value, used to communicate between modules}

\item{output}{Shiny output variable}

\item{error_message}{Error message that will be displayed to user, after translation (character())}

\item{error_name}{Name of the error, to add an entry in the log table (character)}

\item{category}{Category : error or warning ? (character)}

\item{error_report}{Report of the error (character)}

\item{i18n}{Translator object from shiny.i18n library}
}
\description{
Report bug
}
\examples{
\dontrun{
report_bug(r = r, output = output, error_message = "failed_execute_code", 
  error_name = "failed_execute_code", category = "Error", error_report = toString(error_report, i18n = i18n))
}
}
