% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{run_datamart_code}
\alias{run_datamart_code}
\title{Run datamart code}
\usage{
run_datamart_code(
  output,
  r = shiny::reactiveValues(),
  datamart_id = integer(),
  language = "EN",
  quiet = FALSE
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{datamart_id}{ID of datamart containing the code (integer)}

\item{language}{language used for error / warning messages (character, default = "EN")}
}
\description{
Runs datamart code
}
\details{
...
}
\examples{
\dontrun{
run_datamart_code(output = output, r = r, datamart_id = 3)
}
}
