% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{run_dataset_code}
\alias{run_dataset_code}
\title{Run dataset code}
\usage{
run_dataset_code(
  output,
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  dataset_id = integer(),
  i18n = character(),
  quiet = TRUE
)
}
\arguments{
\item{output}{Output variable from Shiny, used to render messages on message bars}

\item{r}{A shiny::reactiValues object, used to communicate between modules}

\item{d}{A shiny::reactiValues object, used to communicate between modules. Contains data loaded from dataset code (d$patients, d$labs_vitals...).}

\item{dataset_id}{ID of dataset containing the code (integer)}

\item{i18n}{shiny.i18n object for translations}
}
\description{
Runs dataset code
}
\details{
The code to load the dataset is loaded from the application database, in code table, from dataset_id.
}
\examples{
\dontrun{
run_dataset_code(output = output, r = r, d = d, i18n = i18n, dataset_id = 3)
}
}
