% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{save_settings_code}
\alias{save_settings_code}
\title{Save code edition}
\usage{
save_settings_code(
  output,
  r = shiny::reactiveValues(),
  id = character(),
  category = character(),
  code_id_input = integer(),
  edited_code = character(),
  i18n = character()
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{category}{Category column in code table, eg : "datamart", "plugin" (character)}

\item{code_id_input}{Input of the actionButton containing ID of current row, in datatable, format = "edit_code_[ID]" (character)}

\item{edited_code}{New code, after editing it (character)}

\item{language}{Language used}
}
\description{
Save code in code table after editing it
}
\examples{
\dontrun{
save_settings_code(output = output, r = r, id = "settings_datamart", category = "datamart", code_id_input = "edit_code_5",
  edited_code = "print('test code edition')", language = "EN")
}
}
