% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{save_settings_options}
\alias{save_settings_options}
\title{Save options}
\usage{
save_settings_options(
  output,
  r = shiny::reactiveValues(),
  id = character(),
  category = character(),
  code_id_input = integer(),
  data = data,
  language = "EN",
  page_options = character()
)
}
\arguments{
\item{output}{variable from Shiny, used to render messages on the message bar}

\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{id}{ID of the current page, format = "settings_[PAGE]" (character)}

\item{category}{Category column in code table, eg : "datamart", "plugin" (character)}

\item{code_id_input}{Input of the actionButton containing ID of current row, in datatable, format = "edit_code_[ID]" (character)}

\item{data}{New data to store in options table (list)}

\item{language}{Language used}
}
\description{
Save options
}
\examples{
\dontrun{
data <- list()
data$show_only_aggregated_data <- TRUE
data$users_allowed_read <- c(1, 3, 4)
save_settings_options(output = output, r = r, id = "settings_datamart", category = "datamart", code_id_input = "edit_code_3",
  data = data, language = "EN")
}
}
