% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{show_message_bar}
\alias{show_message_bar}
\title{Display an error message}
\usage{
show_message_bar(
  output,
  id = integer(),
  message = character(),
  type = "severeWarning",
  language = "EN",
  words = tibble::tibble(),
  time = 7000
)
}
\arguments{
\item{output}{Shiny output variable}

\item{id}{ID of the warning output used (integer)}

\item{message}{message that will be displayed, after translation (character)}

\item{type}{type of message bar displayed (reference to Microsoft MessageBarType enum) (character)}

\item{language}{language used for the translation (character)}

\item{time}{time the message bar will by displayed, in ms (integer)}
}
\description{
Displays an error message on the top of the page
}
\details{
Choose different warning IDs on one page, to allows multiple messages to be displayed.
The different possible types are : c("info", "error", "blocked", "severeWarning", "success", "warning")
}
\examples{
\dontrun{
message_bar(id = 2, message = "name_already_used", type = "severeWarning", language = language, time = 5000)
}
}
