% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{update_r}
\alias{update_r}
\title{Update r variable}
\usage{
update_r(r = shiny::reactiveValues(), table = character(), language = "EN")
}
\arguments{
\item{r}{The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)}

\item{table}{Database table name (character)}

\item{language}{language used for the translation (character)}
}
\description{
Update r value, requesting the corresponding table in the database
}
\examples{
\dontrun{
update_r(r = r, table = "subsets")
}
}
