% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{update_settings_datatable}
\alias{update_settings_datatable}
\title{Update datatable}
\usage{
update_settings_datatable(
  input,
  tab_id = character(),
  r = shiny::reactiveValues(),
  ns = character(),
  table = character(),
  dropdowns = character(),
  i18n = character()
)
}
\arguments{
\item{input}{Shiny input variable}

\item{r}{Shiny r reactive value to communicate between tabs}

\item{ns}{Shiny namespace}

\item{table}{Name of the table used (character)}

\item{dropdowns}{Dropdowns shown on datatable (character)}

\item{language}{Language used (character)}
}
\description{
Update datatable
}
\examples{
\dontrun{
update_settings_datatable(r = r, ns = ns, table = "datasets", dropdowns = "data_source", language = "EN")
}
}
